// src/app/model/thing-to-do/tour-detail/extra-detail-info.ts
export class ExtraDetailInfo {
    id: string;
    adultSlot: number;
    childSlot: number;
    extraPackageId: string;
    fromTime: string;
    toTime: string;
    extraPrice: number;
    extraPriceForChild: number;
    images: string[];
    title: string;
    description: string;
    features: string[];
    activity: string[];
    highlights: string[];
    includes: string[];
    notes: string[];
}